- #43 refactoring get logged in user api
